// frontend/src/api/models/ClientRead.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type ClientRead = {
    full_name: string;
    phone: string;
    email?: (string | null);
    city?: (string | null);
    area?: (string | null);
    relation?: (string | null);
    id: number;
    created_at: string;
    updated_at: string;
};